requirements.txt
flask==3.0.0
torch==2.1.0
torchvision==0.16.0
Pillow==10.1.0
werkzeug==3.0.1
gunicorn==21.2.0